# docs/swagger/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "example api",
    "version": "1.0.0",
    "title": "example api"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/person/{id}": {
      "get": {
        "tags": [
          "person-controller"
        ],
        "summary": "getPerson",
        "operationId": "getPersonUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "delay",
            "in": "query",
            "description": "delay",
            "required": false,
            "type": "integer",
            "default": 2,
            "format": "int32"
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Person"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
...
    "Person": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "Person"
    },
...
  }
}